move new class back as inner class
